--- apps/front-end/src/pages/front-end/ag-edit/upload-photo/schema.js
@@ -1,7 +1,0 @@
-export default {
-  description: "1.Basic details",
-  type: "step",
-  properties: {
-    5: {},
-  },
-};
