--- apps/front-end/src/pages/front-end/ag-edit/upload-photo/uploadphoto.js
@@ -1,594 +1,0 @@
-import React, { useState } from "react";
-import Form from "@rjsf/core";
-import validator from "@rjsf/validator-ajv8";
-import schema1 from "./schema.js";
-import {
-  Alert,
-  Box,
-  Button,
-  Center,
-  HStack,
-  Image,
-  Modal,
-  Radio,
-  Stack,
-  VStack,
-} from "native-base";
-import { CustomRadio } from "@shiksha/common-lib";
-import Steper from "../../../../component/Steper.js";
-import {
-  facilitatorRegistryService,
-  geolocationRegistryService,
-  Camera,
-  Layout,
-  H1,
-  t,
-  login,
-  H3,
-  IconByName,
-  BodySmall,
-  filtersByObject,
-  H2,
-  getBase64,
-  BodyMedium,
-  changeLanguage,
-  enumRegistryService,
-  benificiaryRegistoryService,
-  AgRegistryService,
-  uploadRegistryService,
-  FrontEndTypo,
-} from "@shiksha/common-lib";
-import moment from "moment";
-import { useNavigate, useParams } from "react-router-dom";
-
-import { useScreenshot } from "use-screenshot-hook";
-
-import Clipboard from "component/Clipboard.js";
-import {
-  TitleFieldTemplate,
-  DescriptionFieldTemplate,
-  FieldTemplate,
-  ObjectFieldTemplate,
-  ArrayFieldTitleTemplate,
-  BaseInputTemplate,
-  RadioBtn,
-  CustomR,
-} from "../../../../component/BaseInput.js";
-
-// App
-export default function agFormEdit({ ip }) {
-  const [page, setPage] = React.useState();
-  const [pages, setPages] = React.useState();
-  const [cameraData, setCameraData] = React.useState([]);
-  const [schema, setSchema] = React.useState({});
-  const [cameraSelection, setCameraSelection] = React.useState(0);
-  const [cameraModal, setCameraModal] = React.useState(false);
-  const [credentials, setCredentials] = React.useState();
-  const [cameraUrl, setCameraUrl] = React.useState();
-  const [submitBtn, setSubmitBtn] = React.useState();
-  const [addBtn, setAddBtn] = React.useState(t("YES"));
-  const formRef = React.useRef();
-  const uplodInputRef = React.useRef();
-  const [formData, setFormData] = React.useState({});
-  const [errors, setErrors] = React.useState({});
-  const [alert, setAlert] = React.useState();
-  const [yearsRange, setYearsRange] = React.useState([1980, 2030]);
-  const [lang, setLang] = React.useState(localStorage.getItem("lang"));
-  const { id } = useParams();
-  const [userId, setuserId] = React.useState(id);
-  const navigate = useNavigate();
-
-  const onPressBackButton = async () => {
-    navigate(`/beneficiary/edit/${userId}`);
-  };
-  const ref = React.createRef(null);
-  const { image, takeScreenshot } = useScreenshot();
-  const getImage = () => takeScreenshot({ ref });
-  const downloadImage = () => {
-    var FileSaver = require("file-saver");
-    FileSaver.saveAs(`${image}`, "image.png");
-  };
-
-  React.useEffect(async () => {
-    const qData = await benificiaryRegistoryService.getOne(id);
-    setFormData(qData.result);
-  }, []);
-
-  const nextPreviewStep = async (pageStape = "n") => {
-    setAlert();
-    const index = pages.indexOf(page);
-    const properties = schema1.properties;
-    if (index !== undefined) {
-      let nextIndex = "";
-      if (pageStape.toLowerCase() === "n") {
-        nextIndex = pages[index + 1];
-      } else {
-        nextIndex = pages[index - 1];
-      }
-      if (nextIndex !== undefined) {
-        setPage(nextIndex);
-        setSchema(properties[nextIndex]);
-      } else if (pageStape.toLowerCase() === "n") {
-        await formSubmitUpdate({ ...formData, form_step_number: "6" });
-        setPage("SAVE");
-      } else {
-        return true;
-      }
-    }
-  };
-  const setStep = async (pageNumber = "") => {
-    if (schema1.type === "step") {
-      const properties = schema1.properties;
-      if (pageNumber !== "") {
-        if (page !== pageNumber) {
-          setPage(pageNumber);
-          setSchema(properties[pageNumber]);
-        }
-      } else {
-        nextPreviewStep();
-      }
-    }
-  };
-
-  React.useEffect(() => {
-    if (schema1.type === "step") {
-      const properties = schema1.properties;
-      const newSteps = Object.keys(properties);
-      setPage(newSteps[0]);
-      setSchema(properties[newSteps[0]]);
-      setPages(newSteps);
-      let minYear = moment().subtract("years", 30);
-      let maxYear = moment().subtract("years", 18);
-      setYearsRange([minYear.year(), maxYear.year()]);
-      setSubmitBtn(t("NEXT"));
-    }
-  }, []);
-
-  const formSubmitUpdate = async (formData) => {
-    console.log("sent data");
-    if (id) {
-      const data = await enumRegistryService.editProfileById({
-        ...formData,
-        id: id,
-      });
-      console.log(data, "sent data");
-    }
-  };
-
-  const goErrorPage = (key) => {
-    if (key) {
-      pages.forEach((e) => {
-        console.log(e);
-        const data = schema1["properties"]?.[e]["properties"]?.[key];
-        if (data) {
-          setStep(e);
-        }
-      });
-    }
-  };
-
-  const transformErrors = (errors, uiSchema) => {
-    return errors.map((error) => {
-      if (error.name === "required") {
-        if (schema?.properties?.[error?.property]?.title) {
-          error.message = `${t("REQUIRED_MESSAGE")} "${t(
-            schema?.properties?.[error?.property]?.title
-          )}"`;
-        } else {
-          error.message = `${t("REQUIRED_MESSAGE")}`;
-        }
-      } else if (error.name === "enum") {
-        error.message = `${t("SELECT_MESSAGE")}`;
-      }
-      return error;
-    });
-  };
-
-  const onChange = async (e, id) => {
-    const data = e.formData;
-    setErrors();
-    const newData = { ...formData, ...data };
-    setFormData(newData);
-    if (id === "root_mobile") {
-      if (data?.mobile?.toString()?.length === 10) {
-        const result = await userExist({ mobile: data?.mobile });
-        if (result.isUserExist) {
-          const newErrors = {
-            mobile: {
-              __errors: [t("MOBILE_NUMBER_ALREADY_EXISTS")],
-            },
-          };
-          setErrors(newErrors);
-        }
-      }
-    }
-    if (id === "root_aadhar_token") {
-      if (data?.aadhar_token?.toString()?.length === 12) {
-        const result = await userExist({
-          aadhar_token: data?.aadhar_token,
-        });
-        if (result.isUserExist) {
-          const newErrors = {
-            aadhar_token: {
-              __errors: [t("AADHAAR_NUMBER_ALREADY_EXISTS")],
-            },
-          };
-          setErrors(newErrors);
-        }
-      }
-    }
-
-    if (id === "root_qualification") {
-      if (schema?.properties?.qualification) {
-        let valueIndex = "";
-        schema?.properties?.qualification?.enumNames?.forEach((e, index) => {
-          if (e.match("12")) {
-            valueIndex = schema?.properties?.qualification?.enum[index];
-          }
-        });
-        if (valueIndex !== "" && data.qualification == valueIndex) {
-          setAlert(t("YOU_NOT_ELIGIBLE"));
-        } else {
-          setAlert();
-        }
-      }
-    }
-
-    if (id === "root_device_ownership") {
-      if (schema?.properties?.device_ownership) {
-        if (data?.device_ownership == "no") {
-          setAlert(t("YOU_NOT_ELIGIBLE"));
-        } else {
-          setAlert();
-        }
-      }
-    }
-
-    if (id === "root_state") {
-      await setDistric({
-        schemaData: schema,
-        state: data?.state,
-        district: data?.district,
-        block: data?.block,
-      });
-    }
-
-    if (id === "root_district") {
-      await setBlock({
-        district: data?.district,
-        block: data?.block,
-        schemaData: schema,
-      });
-    }
-
-    if (id === "root_block") {
-      await setVilage({ block: data?.block, schemaData: schema });
-    }
-  };
-
-  const onError = (data) => {
-    console.log(data);
-    if (data[0]) {
-      const key = data[0]?.property?.slice(1);
-      goErrorPage(key);
-    }
-  };
-
-  const onSubmit = async (data) => {
-    const updateDetails = await AgRegistryService.updateAg(formData, userId);
-    console.log("page3.....", updateDetails);
-  };
-
-  const [cameraFile, setcameraFile] = useState();
-
-  const handleFileInputChange = async (e) => {
-    let file = e.target.files[0];
-    if (file.size <= 1048576 * 2) {
-      const data = await getBase64(file);
-      setCameraUrl(data);
-      setcameraFile(file);
-      setFormData({ ...formData, ["profile_url"]: data });
-    } else {
-      setErrors({ fileSize: t("FILE_SIZE") });
-    }
-  };
-
-  const uploadProfile = async () => {
-    // const { id } = authUser;
-    if (userId) {
-      const form_data = new FormData();
-      const item = {
-        file: cameraFile,
-        document_type: "profile",
-        user_id: userId,
-      };
-      for (let key in item) {
-        form_data.append(key, item[key]);
-      }
-
-      const uploadDoc = await uploadRegistryService.uploadFile(form_data);
-      console.log("uploadDoc", uploadDoc);
-    }
-  };
-
-  if (cameraUrl) {
-    return (
-      <Layout
-        _appBar={{
-          lang,
-          setLang,
-          onPressBackButton: (e) => {
-            setCameraUrl();
-            setCameraModal(false);
-          },
-        }}
-        _page={{ _scollView: { bg: "white" } }}
-      >
-        <VStack py={6} px={4} mb={5} space="6">
-          {/* add the profile image */}
-          <FrontEndTypo.H2 color="textMaroonColor.400">
-            {t("ADD_ID_PHOTOS")}
-          </FrontEndTypo.H2>
-          <FrontEndTypo.H4 color="textMaroonColor.400">
-            {t("CLEAR_PROFILE_MESSAGE")}
-          </FrontEndTypo.H4>
-
-          <Center>
-            <Image
-              source={{
-                uri: cameraUrl,
-                //show the image
-              }}
-              alt=""
-              size="324px"
-            />
-          </Center>
-          <FrontEndTypo.H2 color="worksheetBoxText.400">
-            {t("FRONT_VIEW")}
-          </FrontEndTypo.H2>
-          <FrontEndTypo.Primarybutton
-            onPress={async (e) => {
-              if (cameraSelection >= 2) {
-                nextPreviewStep();
-                setCameraUrl();
-                setCameraModal(false);
-                pages[pages?.length - 1] === page;
-              } else {
-                console.log("reached else ==>");
-                uploadProfile();
-                setCameraUrl();
-                setCameraModal(true);
-                setCameraSelection(cameraSelection + 1);
-              }
-              /*  await formSubmitUpdate({ ...formData, form_step_number: "5" });
-              console.log(...formData);
-              if (onClick) onClick(cameraUrl); */
-            }}
-          >
-            {cameraSelection > 2 ? setSubmitBtn(t("SAVE")) : t("SAVE")}
-          </FrontEndTypo.Primarybutton>
-          <FrontEndTypo.Secondarybutton
-            leftIcon={<IconByName name="CameraLineIcon" isDisabled />}
-            onPress={(e) => {
-              setCameraUrl();
-              setCameraModal(true);
-            }}
-          >
-            {t("TAKE_ANOTHER_PHOTO")}
-          </FrontEndTypo.Secondarybutton>
-        </VStack>
-      </Layout>
-    );
-  }
-  if (cameraModal) {
-    return (
-      <Camera
-        {...{
-          cameraModal,
-          setCameraModal,
-          cameraUrl,
-          setcameraFile,
-          setCameraUrl: async (url) => {
-            setCameraUrl(url);
-            setCameraData([...cameraData, url]);
-            setFormData({ ...formData, ["profile_url"]: cameraData });
-          },
-        }}
-      />
-    );
-  }
-
-  if (page === "5") {
-    return (
-      <Layout
-        _appBar={{
-          onPressBackButton,
-          lang,
-          setLang,
-          onlyIconsShow: ["backBtn", "userInfo"],
-        }}
-        _page={{ _scollView: { bg: "white" } }}
-      >
-        <VStack py={6} px={4} mb={5} space="6">
-          <FrontEndTypo.H2 color="textMaroonColor.400">
-            {t("ADD_ID_PHOTOS")}
-          </FrontEndTypo.H2>
-          <FrontEndTypo.H4 color="textMaroonColor.400">
-            {t("CLEAR_PROFILE_MESSAGE")}
-          </FrontEndTypo.H4>
-          <FrontEndTypo.Primarybutton
-            leftIcon={
-              <IconByName
-                name="CameraLineIcon"
-                color="white"
-                size={2}
-                isDisabled
-              />
-            }
-            onPress={(e) => {
-              setCameraUrl();
-              setCameraModal(true);
-            }}
-          >
-            {t("TAKE_PHOTO")}
-          </FrontEndTypo.Primarybutton>
-          <VStack space={2}>
-            <input
-              accept="image/*"
-              type="file"
-              style={{ display: "none" }}
-              ref={uplodInputRef}
-              onChange={handleFileInputChange}
-            />
-            {errors?.fileSize ? (
-              <FrontEndTypo.H2 color="textMaroonColor.400">
-                {errors?.fileSize}
-              </FrontEndTypo.H2>
-            ) : (
-              <React.Fragment />
-            )}
-          </VStack>
-        </VStack>
-      </Layout>
-    );
-  }
-
-  return (
-    <Layout
-      _appBar={{
-        onPressBackButton,
-        name: `${ip?.name}`.trim(),
-        lang,
-        setLang,
-      }}
-      _page={{ _scollView: { bg: "white" } }}
-    >
-      <Box py={6} px={4} mb={5}>
-        {/* Box */}
-        {alert ? (
-          <Alert status="warning" alignItems={"start"} mb="3">
-            <HStack alignItems="center" space="2" color>
-              <Alert.Icon />
-              <BodyMedium>{alert}</BodyMedium>
-            </HStack>
-          </Alert>
-        ) : (
-          <React.Fragment />
-        )}
-        {page && page !== "" ? (
-          <Form
-            key={lang + addBtn}
-            ref={formRef}
-            widgets={{ RadioBtn, CustomR }}
-            templates={{
-              FieldTemplate,
-              ArrayFieldTitleTemplate,
-              ObjectFieldTemplate,
-              TitleFieldTemplate,
-              DescriptionFieldTemplate,
-              BaseInputTemplate,
-            }}
-            extraErrors={errors}
-            showErrorList={false}
-            noHtml5Validate={true}
-            {...{
-              validator,
-              schema: schema ? schema : {},
-              formData,
-              onChange,
-              onError,
-              onSubmit,
-              transformErrors,
-            }}
-          >
-            <FrontEndTypo.Primarybutton
-              mt="3"
-              type="submit"
-              onPress={() => formRef?.current?.submit()}
-            >
-              {pages[pages?.length - 1] === page ? "Submit" : submitBtn}
-            </FrontEndTypo.Primarybutton>
-          </Form>
-        ) : (
-          <React.Fragment />
-        )}
-      </Box>
-      <Modal
-        isOpen={credentials}
-        onClose={() => setCredentials(false)}
-        safeAreaTop={true}
-        size="xl"
-      >
-        <Modal.Content>
-          {/* <Modal.CloseButton /> */}
-          <Modal.Header p="5" borderBottomWidth="0">
-            <FrontEndTypo.H1 textAlign="center">
-              {t("STORE_YOUR_CREDENTIALS")}
-            </FrontEndTypo.H1>
-          </Modal.Header>
-          <Modal.Body p="5" pb="10">
-            <VStack space="5">
-              <VStack alignItems="center">
-                <Box
-                  bg="gray.100"
-                  p="1"
-                  rounded="lg"
-                  borderWidth={1}
-                  borderColor="gray.300"
-                >
-                  <HStack alignItems="center" space="5">
-                    <FrontEndTypo.H3>{t("USERNAME")}</FrontEndTypo.H3>
-                    <FrontEndTypo.H4 ellipsis>
-                      {credentials?.username}
-                    </FrontEndTypo.H4>
-                  </HStack>
-                  <HStack alignItems="center" space="5">
-                    <FrontEndTypo.H3>{t("PASSWORD")}</FrontEndTypo.H3>
-                    <FrontEndTypo.H4 ellipsis>
-                      {credentials?.password}
-                    </FrontEndTypo.H4>
-                  </HStack>
-                </Box>
-              </VStack>
-              <VStack alignItems="center">
-                <Clipboard
-                  text={`username:${credentials?.username}, password:${credentials?.password}`}
-                  onPress={(e) => {
-                    setCredentials({ ...credentials, copy: true });
-                    downloadImage();
-                  }}
-                >
-                  <HStack space="3">
-                    <IconByName
-                      name="FileCopyLineIcon"
-                      isDisabled
-                      rounded="full"
-                      color="blue.300"
-                    />
-                    <FrontEndTypo.H3 color="blue.300">
-                      {t("CLICK_HERE_TO_COPY_AND_LOGIN")}
-                    </FrontEndTypo.H3>
-                  </HStack>
-                </Clipboard>
-              </VStack>
-              <HStack space="5" pt="5">
-                <FrontEndTypo.Primarybutton
-                  flex={1}
-                  isDisabled={!credentials?.copy}
-                  onPress={async (e) => {
-                    const { copy, ...cData } = credentials;
-                    const loginData = await login(cData);
-                    navigate("/");
-                    navigate(0);
-                  }}
-                >
-                  {t("LOGIN")}
-                </FrontEndTypo.Primarybutton>
-              </HStack>
-            </VStack>
-          </Modal.Body>
-        </Modal.Content>
-      </Modal>
-    </Layout>
-  );
-}
