--- apps/front-end/src/pages/front-end/ag-form/parts/Schema.js
@@ -1,51 +1,0 @@
-export default {
-  description: "1.Basic details",
-  type: "step",
-  properties: {
-    1: {
-      title: "IDENTIFY_BENEFICIARY",
-      type: "object",
-      required: ["first_name", "dob"],
-      properties: {
-        first_name: {
-          type: "string",
-          title: "FIRST_NAME",
-          regex: /^(?!.*[\u0900-\u097F])[A-Za-z\s\p{P}]+$/,
-        },
-        middle_name: {
-          type: "string",
-          title: "MIDDLE_NAME",
-          regex: /^(?!.*[\u0900-\u097F])[A-Za-z\s\p{P}]+$/,
-        },
-        last_name: {
-          type: ["string", "null"],
-          title: "LAST_NAME",
-          regex: /^(?!.*[\u0900-\u097F])[A-Za-z\s\p{P}]+$/,
-        },
-        dob: {
-          type: "string",
-          format: "date",
-          label: "DATE_OF_BIRTH_AS_PER_AADHAAR",
-        },
-        role: {
-          format: "hidden",
-          type: "string",
-          default: "beneficiary",
-        },
-      },
-    },
-    2: {
-      title: "CONTACT_INFORMATION",
-      description: "LEARNERS_WHATSAPP_NUMBER",
-      type: "object",
-      required: ["mobile"],
-      properties: {
-        mobile: {
-          type: "string",
-          title: "MOBILE_NUMBER",
-          format: "MobileNumber",
-        },
-      },
-    },
-  },
-};
