+++ apps/front-end/src/v2/components/Functional/LearnerUpdateDetail/contact-details/contactDetailsEdit.js
@@ -25,7 +25,7 @@ import {
   BaseInputTemplate,
   RadioBtn,
   CustomR,
-} from "../../../../component/BaseInput.js";
+} from "../../../Static/FormBaseInput/FormBaseInput.js";
 import { useTranslation } from "react-i18next";
 
 // App
