+++ lib/common-lib/src/services/AgRegistryService.js
@@ -23,6 +23,29 @@ export const createAg = async (formData, header = {}) => {
   }
 }
 
+export const createBeneficiary = async (formData, header = {}) => {
+  try {
+    let headers = {
+      ...header,
+      Authorization: 'Bearer ' + localStorage.getItem('token')
+    }
+    const result = await post(
+      `${process.env.REACT_APP_API_URL}/userauth/register/beneficiary`,
+      formData,
+      {
+        headers: headers
+      }
+    )
+    if (result?.data) {
+      return result?.data
+    } else {
+      return {}
+    }
+  } catch (e) {
+    return handleResponseException(e)
+  }
+}
+
 export const updateAg = async (formData, userId, header = {}) => {
   try {
     let headers = {
